requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
torch==2.1.0
fasttext-wheel==0.9.2
numpy==1.26.0
python-multipart==0.0.6